# src/main/resources/static/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: AWS CardDemo API
  description: "REST API for the AWS CardDemo credit-card account management system,\
    \ migrated from COBOL/CICS to Java 25 and Spring Boot. It re-expresses the 17\
    \ legacy online (3270/BMS) screens as JSON request/response endpoints: signon,\
    \ the main and administrative menus, account view/update, card list/view/update,\
    \ transaction list/view/add, transaction reports, bill payment, and administrative\
    \ user management. This is a faithful, behavior-preserving migration and not a\
    \ web UI, terminal emulator, or 3270 renderer; no capabilities beyond the original\
    \ COBOL application are added."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
- url: /
  description: Relative to the serving host
security:
- basicAuth: []
...
paths:
...
  /api/v1/transactions/view:
    get:
      tags:
      - Transaction View
      summary: Blank Transaction View screen
      description: First-entry screen prompting for a transaction id (COBOL COTRN01C
        initial SEND MAP).
      operationId: initialScreen
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionViewResponse"
        "400":
          description: "Bad Request: the request was malformed or failed input validation\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Unauthorized: authentication is required and was missing or\
            \ invalid (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: "Not Found: the requested resource or record does not exist\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: "Method Not Allowed: the HTTP method is not supported for this\
            \ resource (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: "Internal Server Error: an unexpected error occurred (RFC 7807\
            \ problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
      - Transaction View
      summary: Submit the Transaction View screen
      description: "Reproduces the COTRN01C lookup: fetch and display the requested\
        \ transaction detail."
      operationId: view
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionViewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionViewResponse"
        "400":
          description: "Bad Request: the request was malformed or failed input validation\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Unauthorized: authentication is required and was missing or\
            \ invalid (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: "Not Found: the requested resource or record does not exist\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: "Method Not Allowed: the HTTP method is not supported for this\
            \ resource (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "415":
          description: "Unsupported Media Type: the request body is not application/json\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: "Internal Server Error: an unexpected error occurred (RFC 7807\
            \ problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/accounts/view:
    get:
      tags:
      - Account View
      summary: Blank Account View screen
      description: First-entry screen with the prompt to enter an account id (COBOL
        COACTVWC initial SEND MAP).
      operationId: initialScreen_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountViewResponse"
        "400":
          description: "Bad Request: the request was malformed or failed input validation\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Unauthorized: authentication is required and was missing or\
            \ invalid (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: "Not Found: the requested resource or record does not exist\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: "Method Not Allowed: the HTTP method is not supported for this\
            \ resource (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: "Internal Server Error: an unexpected error occurred (RFC 7807\
            \ problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
      - Account View
      summary: Submit the Account View screen
      description: "Routes on the transmitted key: ENTER fetches the account, PF3\
        \ navigates back, any other key redisplays with an invalid-key message."
      operationId: view_3
      parameters:
      - name: X-CardDemo-From-Program
        in: header
        required: false
        schema:
          type: string
      - name: X-CardDemo-From-Transaction
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountViewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountViewResponse"
        "400":
          description: "Bad Request: the request was malformed or failed input validation\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Unauthorized: authentication is required and was missing or\
            \ invalid (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: "Not Found: the requested resource or record does not exist\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: "Method Not Allowed: the HTTP method is not supported for this\
            \ resource (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "415":
          description: "Unsupported Media Type: the request body is not application/json\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: "Internal Server Error: an unexpected error occurred (RFC 7807\
            \ problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    TransactionViewRequest:
      type: object
      properties:
        transactionId:
          type: string
          maxLength: 16
          minLength: 0
        action:
          type: string
          enum:
          - ENTER
          - CLEAR
          - PA1
          - PA2
          - PF1
          - PF2
          - PF3
          - PF4
          - PF5
          - PF6
          - PF7
          - PF8
          - PF9
          - PF10
          - PF11
          - PF12
    TransactionViewResponse:
      type: object
      properties:
        transactionName:
          type: string
        title01:
          type: string
        title02:
          type: string
        currentDate:
          type: string
        programName:
          type: string
        currentTime:
          type: string
        transactionId:
          type: string
        cardNumber:
          type: string
        typeCode:
          type: string
        categoryCode:
          type: string
        source:
          type: string
        description:
          type: string
        amount:
          type: number
        originDate:
          type: string
        processDate:
          type: string
        merchantId:
          type: string
        merchantName:
          type: string
        merchantCity:
          type: string
        merchantZip:
          type: string
        errorMessage:
          type: string
...
    AccountViewRequest:
      type: object
      properties:
        accountId:
          type: string
          maxLength: 11
          minLength: 0
          pattern: "^\\d{1,11}$"
        action:
          type: string
          enum:
          - ENTER
          - CLEAR
          - PA1
          - PA2
          - PF1
          - PF2
          - PF3
          - PF4
          - PF5
          - PF6
          - PF7
          - PF8
          - PF9
          - PF10
          - PF11
          - PF12
      required:
      - accountId
    AccountViewResponse:
      type: object
      properties:
        transactionName:
          type: string
        title01:
          type: string
        currentDate:
          type: string
        programName:
          type: string
        title02:
          type: string
        currentTime:
          type: string
        accountId:
          type: string
        accountStatus:
          type: string
        dateOpened:
          type: string
        creditLimit:
          type: number
        expiryDate:
          type: string
        cashLimit:
          type: number
        reissueDate:
          type: string
        currentBalance:
          type: number
        currentCycleCredit:
          type: number
        groupId:
          type: string
        currentCycleDebit:
          type: number
        customerId:
          type: string
        ssn:
          type: string
        dateOfBirth:
          type: string
        ficoScore:
          type: string
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        addressLine1:
          type: string
        stateCode:
          type: string
        addressLine2:
          type: string
        zipCode:
          type: string
        city:
          type: string
        countryCode:
          type: string
        phone1:
          type: string
        governmentId:
          type: string
        phone2:
          type: string
        eftAccountId:
          type: string
        primaryHolderFlag:
          type: string
        infoMessage:
          type: string
        errorMessage:
          type: string
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem detail returned as application/problem+json for
        every error response emitted by the API.
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          format: int32
          example: 404
        detail:
          type: string
          example: The requested account was not found.
        instance:
          type: string
          format: uri
          example: /api/v1/accounts/view
        correlationId:
          type: string
          description: Request correlation id echoed from the X-Correlation-Id header
            / MDC.
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic